evaluation/tools/SCRepair/screpair/requirements.txt
py-solc==3.2.0
Logbook==1.4.1
deap==1.3.0
attrs==18.2.0
docker==3.7.2
numpy==1.16.2
matplotlib==3.4.1
